requirements.txt
# Agno Document Q&A System Requirements
# Core AI and ML frameworks
agno>=0.1.0
groq>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0

# Web interface
streamlit>=1.28.0

# Environment and utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Data handling
pandas>=1.5.0
json5>=0.9.0

# Optional: For better performance
torch>=2.0.0
transformers>=4.21.0

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
